Refactor adapter test modules
