+++ tests_adapters/models/test_clip.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from tests.models.clip.test_modeling_clip import *  # Imported to execute model tests
