+++ tests_adapters/models/test_encoder_decoder.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from tests.models.encoder_decoder.test_modeling_encoder_decoder import *  # Imported to execute model tests
