+++ tests_adapters/models/test_gptj.py
@@ -1,0 +1,11 @@
+from tests.models.gptj.test_modeling_gptj import *
+from transformers import GPTJAdapterModel
+from transformers.testing_utils import require_torch
+
+from .base import AdapterModelTesterMixin
+
+
+@require_torch
+class GPTJAdapterModelTest(AdapterModelTesterMixin, GPTJModelTest):
+    all_model_classes = (GPTJAdapterModel,)
+    fx_compatible = False
