+++ tests_adapters/models/test_vit.py
@@ -1,0 +1,11 @@
+from tests.models.vit.test_modeling_vit import *
+from transformers import ViTAdapterModel
+from transformers.testing_utils import require_torch
+
+from .base import AdapterModelTesterMixin
+
+
+@require_torch
+class ViTAdapterModelTest(AdapterModelTesterMixin, ViTModelTest):
+    all_model_classes = (ViTAdapterModel,)
+    fx_compatible = False
