+++ tests_adapters/models/test_xlm_roberta.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from tests.models.xlm_roberta.test_modeling_xlm_roberta import *  # Imported to execute model tests
