+++ tests_adapters/test_bart.py
@@ -1,7 +1,6 @@
 import unittest
 
-from tests.models.bart.test_modeling_bart import *
-from transformers import BartAdapterModel
+from transformers import BartConfig
 from transformers.testing_utils import require_torch
 
 from .methods import (
@@ -19,15 +18,6 @@ from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class BartAdapterModelTest(AdapterModelTesterMixin, BartModelTest):
-    all_model_classes = (
-        BartAdapterModel,
-    )
-    fx_compatible = False
 
 
 class BartAdapterTestBase(AdapterTestBase):
