+++ tests_adapters/test_beit.py
@@ -1,16 +1,15 @@
 import unittest
 
-from tests.models.beit.test_modeling_beit import *
-from transformers import BeitAdapterModel
+from transformers import BeitConfig
 from transformers.testing_utils import require_torch
 
 from .methods import (
-   BottleneckAdapterTestMixin,
-   CompacterTestMixin,
-   IA3TestMixin,
-   LoRATestMixin,
-   PrefixTuningTestMixin,
-   UniPELTTestMixin,
+    BottleneckAdapterTestMixin,
+    CompacterTestMixin,
+    IA3TestMixin,
+    LoRATestMixin,
+    PrefixTuningTestMixin,
+    UniPELTTestMixin,
 )
 from .test_adapter import VisionAdapterTestBase, make_config
 from .test_adapter_backward_compability import CompabilityTestMixin
@@ -19,15 +18,6 @@ from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class BeitAdapterModelTest(AdapterModelTesterMixin, BeitModelTest):
-    all_model_classes = (
-        BeitAdapterModel,
-    )
-    fx_compatible = False
 
 
 class BeitAdapterTestBase(VisionAdapterTestBase):
@@ -40,7 +30,7 @@ class BeitAdapterTestBase(VisionAdapterTestBase):
         num_attention_heads=4,
         intermediate_size=37,
     )
-    feature_extractor_name = 'microsoft/beit-base-patch16-224-pt22k'
+    feature_extractor_name = "microsoft/beit-base-patch16-224-pt22k"
 
 
 @require_torch
