+++ tests_adapters/test_distilbert.py
@@ -1,10 +1,16 @@
 import unittest
 
-from tests.models.distilbert.test_modeling_distilbert import *
-from transformers import DistilBertAdapterModel
+from transformers import DistilBertConfig
 from transformers.testing_utils import require_torch
 
-from .methods import BottleneckAdapterTestMixin, UniPELTTestMixin, CompacterTestMixin, IA3TestMixin, LoRATestMixin, PrefixTuningTestMixin
+from .methods import (
+    BottleneckAdapterTestMixin,
+    UniPELTTestMixin,
+    CompacterTestMixin,
+    IA3TestMixin,
+    LoRATestMixin,
+    PrefixTuningTestMixin,
+)
 from .test_adapter import AdapterTestBase, make_config
 from .test_adapter_backward_compability import CompabilityTestMixin
 from .test_adapter_composition import ParallelAdapterInferenceTestMixin, ParallelTrainingMixin
@@ -12,15 +18,6 @@ from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class DistilBertAdapterModelTest(AdapterModelTesterMixin, DistilBertModelTest):
-    all_model_classes = (
-        DistilBertAdapterModel,
-    )
-    fx_compatible = False
 
 
 class DistilBertAdapterTestBase(AdapterTestBase):
