+++ tests_adapters/test_gpt2.py
@@ -1,7 +1,6 @@
 import unittest
 
-from tests.models.gpt2.test_modeling_gpt2 import *
-from transformers import GPT2AdapterModel
+from transformers import GPT2Config
 from transformers.testing_utils import require_torch
 
 from .methods import (
@@ -19,15 +18,6 @@ from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class GPT2AdapterModelTest(AdapterModelTesterMixin, GPT2ModelTest):
-    all_model_classes = (
-        GPT2AdapterModel,
-    )
-    fx_compatible = False
 
 
 class GPT2AdapterTestBase(AdapterTestBase):
