+++ tests_adapters/test_mbart.py
@@ -1,24 +1,21 @@
 import unittest
 
-from tests.models.mbart.test_modeling_mbart import *
-from transformers import MBartAdapterModel
+from transformers import MBartConfig
 from transformers.testing_utils import require_torch
 
-from .methods import BottleneckAdapterTestMixin, UniPELTTestMixin, CompacterTestMixin, IA3TestMixin, LoRATestMixin, PrefixTuningTestMixin
+from .methods import (
+    BottleneckAdapterTestMixin,
+    UniPELTTestMixin,
+    CompacterTestMixin,
+    IA3TestMixin,
+    LoRATestMixin,
+    PrefixTuningTestMixin,
+)
 from .test_adapter import AdapterTestBase, make_config
 from .test_adapter_composition import ParallelAdapterInferenceTestMixin
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class MBartAdapterModelTest(AdapterModelTesterMixin, MBartModelTest):
-    all_model_classes = (
-        MBartAdapterModel,
-    )
-    fx_compatible = False
 
 
 class MBartAdapterTestBase(AdapterTestBase):
