+++ tests_adapters/test_roberta.py
@@ -1,25 +1,22 @@
 import unittest
 
-from tests.models.roberta.test_modeling_roberta import *
-from transformers import RobertaAdapterModel
+from transformers import RobertaConfig
 from transformers.testing_utils import require_torch
 
-from .methods import BottleneckAdapterTestMixin, UniPELTTestMixin, CompacterTestMixin, IA3TestMixin, LoRATestMixin, PrefixTuningTestMixin
+from .methods import (
+    BottleneckAdapterTestMixin,
+    UniPELTTestMixin,
+    CompacterTestMixin,
+    IA3TestMixin,
+    LoRATestMixin,
+    PrefixTuningTestMixin,
+)
 from .test_adapter import AdapterTestBase, make_config
 from .test_adapter_backward_compability import CompabilityTestMixin
 from .test_adapter_composition import ParallelAdapterInferenceTestMixin
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class RobertaAdapterModelTest(AdapterModelTesterMixin, RobertaModelTest):
-    all_model_classes = (
-        RobertaAdapterModel,
-    )
-    fx_compatible = False
 
 
 class RobertaAdapterTestBase(AdapterTestBase):
