+++ tests_adapters/test_t5.py
@@ -2,8 +2,7 @@ import unittest
 
 from datasets import load_dataset
 
-from tests.models.t5.test_modeling_t5 import *
-from transformers import AutoTokenizer, T5AdapterModel
+from transformers import AutoTokenizer, T5Config
 from transformers.testing_utils import require_torch
 
 from .methods import (
@@ -21,15 +20,6 @@ from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_common import AdapterModelTesterMixin
-
-
-@require_torch
-class T5AdapterModelTest(AdapterModelTesterMixin, T5ModelTest):
-    all_model_classes = (
-        T5AdapterModel,
-    )
-    fx_compatible = False
 
 
 @require_torch
